backend/requirements.txt
Django~=4.2.7
networkx~=3.4.2
pm4py~=2.7.12.3
djangorestframework~=3.14.0
boto3~=1.35.71
celery~=5.4.0